Fixing external package usage for Python 3.7
When trying the example from the module docs page, I got an ImportError for the quote_plus name and an AttributeError. This commit fixes the ImportError by pointing the import to urllib.parse if applicable, where the quote_plus name is located from Python 3.4 upwards. The AttributeError is addressed by importing the JWT class from the jwt module, and passing an object of the JWT class as the self argument for JWT.decode() instead of the auth token. The auth token should be the second positional argument, corresponding to the "message" parameter of decode().